requirements.txt
streamlit
azure-storage-blob
azure-search-documents
azure-ai-formrecognizer
PyPDF2
requests
transformers
torch
